requirements.txt
aioamqp>=0.13.0
attrs>=19.1.0
sqlalchemy>=1.3.15
alembic>=1.4.2
